backend/requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
--no-binary psycopg2

avro-python3==1.9.1       # via confluent-kafka
boltons==19.2.0
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
confluent-kafka[avro]==1.2.0
django==2.0.13
django-cors-headers==3.1.1
django-enumfields==1.0.0
django-environ==0.4.5
django-extensions==2.2.5
django-helusers==0.5.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-enum-field==0.9.3
drf-oidc-auth==0.9        # via django-helusers
ecdsa==0.13.3             # via python-jose
fastavro==0.22.5          # via confluent-kafka
future==0.18.1            # via pyjwkest, python-jose
idna==2.8                 # via requests
psycopg2==2.8.4
pyasn1==0.4.7             # via rsa
pycryptodomex==3.9.0      # via pyjwkest
pyjwkest==1.4.2           # via drf-oidc-auth
pyjwt==1.7.1              # via djangorestframework-jwt
python-jose==3.0.1        # via django-helusers
pytz==2019.3              # via django
raven==6.10.0
requests==2.22.0
rsa==4.0                  # via python-jose
six==1.12.0               # via django-extensions, pyjwkest, python-jose
urllib3==1.25.6           # via requests
uwsgi==2.0.18
